sub: don't truncate timestamps to int
Realistically probably no problem at all since they're in ms and usually
start from 0, but whatever.

+++ sub/dec_sub.c
@@ -344,7 +344,7 @@ static void multiply_timings(struct packet_list *subs, double factor)
     }
 }
 
-#define MS_TS(f_ts) ((int)((f_ts) * 1000 + 0.5))
+#define MS_TS(f_ts) ((long long)((f_ts) * 1000 + 0.5))
 
 // Remove overlaps and fill gaps between adjacent subtitle packets. This is done
 // by adjusting the duration of the earlier packet. If the gaps or overlap are
